+++ CHANGES.md
@@ -18,7 +18,7 @@ Translations 🗣:
  -
 
 SDK API changes ⚠️:
- - 
+ - Rename `tryThis` to `tryOrNull`
 
 Build 🧱:
  -